Add some basic OpenBSD support, very similar to FreeBSD.
It has some limitations because of the different PHP versions
available on packages mirrors and selecting the right version
you want. It is because the packages are suffixed with the
version, and unfortunately not prefixed, like ruby packages.

So, things work for the default php version available in
packages, but when you want some other php version, you
have to pin the version explicitly so far.

extensions installation management works, php_fpm also
seems to be good ;)

Add $ext_tool_enabled into openbsd case

Add a $config_root_inifile for OpenBSD

Other name for php-cli ini file

switch config_root_inifile with the other config file for cli